// src/types/ContainerChildren.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Container } from "./Container";
import type { StorySummary } from "./StorySummary";

/**
 * Response type containing child containers and stories
 */
export type ContainerChildren = { containers: Array<Container>, stories: Array<StorySummary>, };